move sensors out of external lib folder
